[LISP] More efficient LISP interpreter.
This interpreter derives from the other, but it uses the rest of the
datatypes supported from alloc.cpp. There is some provision in the
read routine for more syntax but it isn't used.

Symbols are not interned. Implementing LISP is a little detour on the
way to implementing Haskell-like and ML-like languages, so interning
symbols will be left to later, if at all.

This version of the interpreter has to work much harder to blow up. So
far the expect script does not seem to drive it hard enough to run out
of heap.