[GC] Adding array and built-in types
Adding the array and built-in function as data types.

Elevating the idea that integers in the heap are often headers, but
don't have to be. So before, an integer in the heap (BOXSTR_INT) was
assumed to be the header for the contents of a string. Now such an
integer will be tagged with BOXSTR_STRHDR. Likewise the array has
its own header type BOXSTR_ARRAYHDR.

Adding this has neither improved nor degraded the behavior of
gcstringlisp.