[GC] Bug fixes and interned symbols.
alloc.cpp/alloc.h:
Removed "TYPE" from names of types.
Added BOXSTR_SYMBOL, an interned string. If a string and symbol are
compared they are compared character for character. This may change.
A Rooter object allows clients to use RAII to define a GC root that
expires when it goes out of scope.

alloc.cpp:
Added an experimental LOG template which is a variadic function
template. Replaced a subset of cout << foo expressions with LOG calls.
In string_append_char and array_append_element, the logic for checking
whether the collection could be appended to without moving it and other
situations was not complete, so arrays could end up extended past the
end of heap.
Forwarding wasn't completely carried out and so items that were
referenced by multiple roots were ending up duplicated.