[LISP] Using symbols and array for env.
Changed the names of the cryptically named functions.
Converted the environment from an STL array of STL pairs to LISP
array of cons pairs. The CAR is a symbol and the CDR is either a
built-in function or an s-expression defining a user's function.
Modified the read_obj routine to return symbols when appropriate.
Added test usage to make it easier to test with GDB.